Make keyexpimp independent test.
Move test code from gost_keyexpimp.c into separate test_keyexpimp.c and
link into test target.

Also, make tests work without requiring config.
